requirements.txt
google_api_python_client==2.172.0
gspread==6.2.1
gspread_dataframe==4.0.0
pandas==2.3.0
protobuf==6.31.1
python-dotenv==1.1.0
Requests==2.32.4
